refactor: List<Car> 를 일급컬렉션 Cars로 변경
